Improve description
