+++ src/main/java/org/skriptlang/skript/bukkit/entity/EntityModule.java
@@ -7,6 +7,7 @@ import org.skriptlang.skript.addon.AddonModule;
 import org.skriptlang.skript.addon.HierarchicalAddonModule;
 import org.skriptlang.skript.addon.SkriptAddon;
 import org.skriptlang.skript.bukkit.entity.displays.DisplayModule;
+import org.skriptlang.skript.bukkit.entity.elements.conditions.CondAI;
 import org.skriptlang.skript.bukkit.entity.elements.conditions.CondHasGravity;
 import org.skriptlang.skript.bukkit.entity.elements.conditions.CondIsGlowing;
 import org.skriptlang.skript.bukkit.entity.elements.conditions.CondProjectileIsCritical;
@@ -45,6 +46,7 @@ public class EntityModule extends HierarchicalAddonModule {
 		}
 
 		register(addon,
+			CondAI::register,
 			CondHasGravity::register,
 			CondIsGlowing::register,
 			CondProjectileIsCritical::register,
